modifiche di pulizia 2

+++ AlgoInC++/codice/KMinHash.h
@@ -23,5 +23,5 @@ public:
      * @param s: Vettore di input.
      * @return Firma k-minHash come vettore di interi.
      */
-    std::vector<double> computeSignature(std::vector<uint64_t> s);
+    std::vector<uint64_t> computeSignature(std::vector<uint64_t> s);
 };
